+++ src/index.js
@@ -2,4 +2,5 @@ export { default } from './extendRoutine';
 
 export { default as createExtendedRoutine } from './createExtendedRoutine';
 export { default as createCustomRoutine } from './createCustomRoutine';
-export { default as createSocketRoutine } from './createSocketRoutine';
+export { default as createSocketRoutine, defaultSocketStages } from './createSocketRoutine';
+export { default as createRoutines } from './createRoutines';