+++ CHANGELOG.md
@@ -24,3 +24,9 @@ note: from this ver you must have at least `redux-saga-routines 3.2.0` installed
 ### Version 3.1.1
 - change default stages for `creteSocketRoutine`:
   - rename 'MESSAGE_SENDED' and 'MESSAGE_RECEIVED' to 'SENDED' and 'RECEIVED'
+
+### Version 3.2.0
+- Change `creteSocketRoutine` default stages to 'CONNECTED', 'DISCONNECTED', 'JOIN_CHANNEL', 'CHANNEL_JOINED', 'LEAVE_CHANNEL', 'CHANNEL_LEAVED' and export defaultSocketStages
+- Ability to create a bunch of routines with new method `createRoutines`
+- Allow to pass one stage as a string for `createCustomRoutine`
+- Fix typo in readme